# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Api",
        "description": "API do projeto Base",
        "contact": {
            "email": "email@test.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base API"
        }
    ],
    "paths": {
...
        "/solicitar": {
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Solicitar acesso",
                "operationId": "solicitar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "name",
                                    "email",
                                    "perfil_id",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "00000000191"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "teste"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste@mail.com"
                                    },
                                    "perfil_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "teste123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "teste123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                }
            }
        },
        "/users": {
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Criar usuário",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "name",
                                    "email",
                                    "perfil_id",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "00000000191"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "teste"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste@mail.com"
                                    },
                                    "perfil_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "teste123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "teste123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/approve/{cpf}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Aprovar usuário",
                "operationId": "aprovarUsuario",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "Cpf do usuário",
                        "schema": {
                            "type": "string",
                            "format": "int64",
                            "example": "00000000191"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/block/{cpf}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Bloquear usuário",
                "operationId": "blockUsuario",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "Cpf do usuário",
                        "schema": {
                            "type": "string",
                            "format": "int64",
                            "example": "00000000191"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/api/auth/auth",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}